django_backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-telegram-bot==22.3
celery==5.5.3
redis==6.2.0
Pillow==10.4.0
reportlab==4.1.0
python-decouple==3.8
urllib3<2.0.0 